Why double!?

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robot/LeoOne.java
@@ -215,11 +215,11 @@ public class LeoOne extends Robot {
         }
     }
 
-    public void moveClaw(int direction) {
+    public void moveClaw(double direction) {
         claw.setPosition(direction);
     }
 
-    public void moveWrist(int direction) {
+    public void moveWrist(double direction) {
         wrist.setPower(direction);
     }
